move funcao para novo modulo
